Fix windows linter complaints
