+++ cmd/nerdctl/container/container_cp_unix.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 /*
    Copyright The containerd Authors.
