+++ cmd/nerdctl/container/container_run_network_linux_test.go
@@ -37,6 +37,31 @@ import (
 	"github.com/containerd/nerdctl/v2/pkg/testutil/nettestutil"
 )
 
+func extractHostPort(portMapping string, port string) (string, error) {
+	// Regular expression to extract host port from port mapping information
+	re := regexp.MustCompile(`(?P<containerPort>\d{1,5})/tcp ->.*?0.0.0.0:(?P<hostPort>\d{1,5}).*?`)
+	portMappingLines := strings.Split(portMapping, "\n")
+	for _, portMappingLine := range portMappingLines {
+		// Find the matches
+		matches := re.FindStringSubmatch(portMappingLine)
+		// Check if there is a match
+		if len(matches) >= 3 && matches[1] == port {
+			// Extract the host port number
+			hostPort := matches[2]
+			return hostPort, nil
+		}
+	}
+	return "", fmt.Errorf("could not extract host port from port mapping: %s", portMapping)
+}
+
+func valuesOfMapStringString(m map[string]string) map[string]struct{} {
+	res := make(map[string]struct{})
+	for _, v := range m {
+		res[v] = struct{}{}
+	}
+	return res
+}
+
 // TestRunInternetConnectivity tests Internet connectivity with `apk update`
 func TestRunInternetConnectivity(t *testing.T) {
 	base := testutil.NewBase(t)
